pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "smart-insurance-ai"
version = "0.1.0"
description = "AI-powered smart insurance platform for inclusive micro-insurance (ML + DeepONet uncertainty)."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name = "سالمة Team" }
]
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "pydantic>=2.0",
  "python-dotenv>=1.0",
  "sqlalchemy>=2.0",


  "streamlit>=1.30",

  "pandas>=2.0",
  "numpy>=1.24",
  "scikit-learn>=1.3",
  "joblib>=1.3",

  "torch>=2.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0",
  "black>=24.0",
  "ruff>=0.5.0"
]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 120
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.flake8]
max-line-length = 120
extend-ignore = ["E501"]
